[tests] Update more tests to use explicit native type comparisons.

+++ tests/monotouch-test/MetalPerformanceShaders/KernelTest.cs
@@ -52,13 +52,13 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 			
 			var r = MPSKernel.RectNoClip;
 			var o = r.Origin;
-			Assert.That (o.X, Is.EqualTo (0), "X");
-			Assert.That (o.Y, Is.EqualTo (0), "Y");
-			Assert.That (o.Z, Is.EqualTo (0), "Z");
+			Assert.That (o.X, Is.EqualTo ((nint) 0), "X");
+			Assert.That (o.Y, Is.EqualTo ((nint) 0), "Y");
+			Assert.That (o.Z, Is.EqualTo ((nint) 0), "Z");
 			var s = r.Size;
-			Assert.That (s.Depth, Is.EqualTo (-1), "Depth");
-			Assert.That (s.Height, Is.EqualTo (-1), "Height");
-			Assert.That (s.Width, Is.EqualTo (-1), "Width");
+			Assert.That (s.Depth, Is.EqualTo ((nint) (-1)), "Depth");
+			Assert.That (s.Height, Is.EqualTo ((nint) (-1)), "Height");
+			Assert.That (s.Width, Is.EqualTo ((nint) (-1)), "Width");
 		}
 
 		[Test]
@@ -72,7 +72,7 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 
 			var kernel = new MPSKernel (MTLDevice.SystemDefault);
 			var kernel2 = kernel.CopyWithZone (NSZone.Default, MTLDevice.SystemDefault);
-			Assert.That (kernel2.RetainCount, Is.EqualTo (1));
+			Assert.That (kernel2.RetainCount, Is.EqualTo ((nuint) 1));
 		}
 
 		[Test]
@@ -86,7 +86,7 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 
 			var layer = new MPSRnnImageInferenceLayer (MTLDevice.SystemDefault, MPSRnnSingleGateDescriptor.Create (1, 1));
 			var layer2 = layer.Copy (NSZone.Default, MTLDevice.SystemDefault);
-			Assert.That (layer2.RetainCount, Is.EqualTo (1));
+			Assert.That (layer2.RetainCount, Is.EqualTo ((nuint) 1));
 		}
 
 		[Test]
@@ -100,7 +100,7 @@ namespace MonoTouchFixtures.MetalPerformanceShaders {
 
 			var layer = new MPSRnnMatrixInferenceLayer (MTLDevice.SystemDefault, MPSRnnSingleGateDescriptor.Create (1, 1));
 			var layer2 = layer.Copy (NSZone.Default, MTLDevice.SystemDefault);
-			Assert.That (layer2.RetainCount, Is.EqualTo (1));
+			Assert.That (layer2.RetainCount, Is.EqualTo ((nuint) 1));
 		}
 
 		[Test]
